fix the type of the dataOffsetTmp

+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuRowToColumnarExec.scala
@@ -857,7 +857,7 @@ object GeneratedInternalRowToCudfRowIterator extends Logging {
          |    Object $rowBaseObj = input.getBaseObject();
          |    long $rowBaseOffset = input.getBaseOffset();
          |
-         |    long $cudfDataOffsetTmp;
+         |    int $cudfDataOffsetTmp;
          |    ${cudfDataOffsetInit}
          |
          |    ${copyData.mkString("\n")}
